Implement entire url creation in pagination adapter
